refactor(feed): #67 - 중복 svg 제거 및 common 폴더로 통합
