+++ assets/svgs/common/comment-icon.js
