--- assets/svgs/common/comment.js
@@ -1,18 +1,0 @@
-import Svg, { Path } from 'react-native-svg';
-
-export default function Comment({ size, color }) {
-  return (
-    <Svg
-      width={size + 1}
-      height={size}
-      viewBox="0 0 14 13"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <Path
-        d="M3.33333 8H10C10.1889 8 10.3473 7.936 10.4753 7.808C10.6033 7.68 10.6671 7.52178 10.6667 7.33333C10.6662 7.14489 10.6022 6.98667 10.4747 6.85867C10.3471 6.73067 10.1889 6.66667 10 6.66667H3.33333C3.14444 6.66667 2.98622 6.73067 2.85867 6.85867C2.73111 6.98667 2.66711 7.14489 2.66667 7.33333C2.66622 7.52178 2.73022 7.68022 2.85867 7.80867C2.98711 7.93711 3.14533 8.00089 3.33333 8ZM3.33333 6H10C10.1889 6 10.3473 5.936 10.4753 5.808C10.6033 5.68 10.6671 5.52178 10.6667 5.33333C10.6662 5.14489 10.6022 4.98667 10.4747 4.85867C10.3471 4.73067 10.1889 4.66667 10 4.66667H3.33333C3.14444 4.66667 2.98622 4.73067 2.85867 4.85867C2.73111 4.98667 2.66711 5.14489 2.66667 5.33333C2.66622 5.52178 2.73022 5.68022 2.85867 5.80867C2.98711 5.93711 3.14533 6.00089 3.33333 6ZM3.33333 4H10C10.1889 4 10.3473 3.936 10.4753 3.808C10.6033 3.68 10.6671 3.52178 10.6667 3.33333C10.6662 3.14489 10.6022 2.98667 10.4747 2.85867C10.3471 2.73067 10.1889 2.66667 10 2.66667H3.33333C3.14444 2.66667 2.98622 2.73067 2.85867 2.85867C2.73111 2.98667 2.66711 3.14489 2.66667 3.33333C2.66622 3.52178 2.73022 3.68022 2.85867 3.80867C2.98711 3.93711 3.14533 4.00089 3.33333 4ZM1.33333 10.6667C0.966667 10.6667 0.652889 10.5362 0.392 10.2753C0.131111 10.0144 0.000444444 9.70045 0 9.33333V1.33333C0 0.966667 0.130667 0.652889 0.392 0.392C0.653333 0.131111 0.967111 0.000444444 1.33333 0H12C12.3667 0 12.6807 0.130667 12.942 0.392C13.2033 0.653333 13.3338 0.967111 13.3333 1.33333V11.7167C13.3333 12.0167 13.1973 12.2251 12.9253 12.342C12.6533 12.4589 12.4116 12.4116 12.2 12.2L10.6667 10.6667H1.33333Z"
-        fill={color}
-      />
-    </Svg>
-  );
-}
