+++ assets/svgs/common/like-icon.js
