--- assets/svgs/common/like.js
@@ -1,18 +1,0 @@
-import Svg, { Path } from 'react-native-svg';
-
-export default function Like({ size, color }) {
-  return (
-    <Svg
-      width={size}
-      height={size}
-      viewBox="0 0 12 12"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <Path
-        d="M2.99375 0.056792C3.52455 -0.0436783 4.06888 -0.0105356 4.58662 0.153777C5.10435 0.31809 5.58229 0.609383 5.98521 1.00619L6.0074 1.02817L6.02779 1.00818C6.41234 0.633288 6.86443 0.353799 7.3539 0.188363C7.84338 0.0229268 8.35898 -0.0246532 8.86632 0.0487973L9.01385 0.0727819C9.65335 0.195443 10.2511 0.507941 10.7438 0.977184C11.2364 1.44643 11.6057 2.05495 11.8125 2.7383C12.0192 3.42166 12.0558 4.15441 11.9183 4.85897C11.7808 5.56353 11.4743 6.21366 11.0314 6.74052L10.9234 6.86378L10.8946 6.89109L6.42661 11.8073C6.32351 11.9207 6.1869 11.9887 6.04212 11.9987C5.89733 12.0087 5.75417 11.9602 5.63917 11.8619L5.58279 11.8073L1.08901 6.86244C0.612951 6.34788 0.27439 5.69698 0.110819 4.98182C-0.0527521 4.26667 -0.0349506 3.51518 0.162252 2.81058C0.359455 2.10597 0.728363 1.47575 1.22812 0.98971C1.72788 0.50367 2.33898 0.180777 2.99375 0.056792Z"
-        fill={color}
-      />
-    </Svg>
-  );
-}
